pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "docforge"
version = "0.1.0"
description = "PDF/文档智能处理 CLI 工具 - 纯本地运行，隐私安全"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "小夏 (OpenClaw)", email = "xiaoxia008@proton.me"}
]
keywords = ["pdf", "cli", "document", "privacy", "local"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "pypdf>=3.0",
    "reportlab>=4.0",
    "rich>=12.0",
    "Pillow>=9.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10", "pdf2image>=1.16"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
docforge = "docforge.cli:cli"

[project.urls]
Homepage = "https://github.com/xiaoxia008/docforge"
Repository = "https://github.com/xiaoxia008/docforge"
Issues = "https://github.com/xiaoxia008/docforge/issues"
